+++ src/decimal/dec/parse/from_float.rs
@@ -1,75 +1,14 @@
+mod f2dec;
+
 use crate::{
     decimal::{
         dec::{ControlBlock, ExtraPrecision},
         signals::Signals,
         Context, Decimal, Sign,
     },
-    int::{math::ilog10, UInt},
 };
 
 type D<const N: usize> = Decimal<N>;
-
-// @TODO Performance optimizations
-
-#[inline]
-const fn f2dec<const N: usize>(mant: u64, b_exp: i16, sign: Sign) -> D<N> {
-    if b_exp == 0 {
-        return D::new(
-            UInt::from_digit(mant),
-            ControlBlock::new(
-                0,
-                sign,
-                Signals::empty(),
-                Context::default(),
-                ExtraPrecision::new(),
-            ),
-        );
-    }
-
-    let uint = if b_exp < 0 {
-        UInt::TWO.pow(-b_exp as u32)
-    } else {
-        UInt::TWO.pow(b_exp as u32)
-    };
-
-    let d_exp = ilog10(uint) as i16 + 1;
-    let psi = D::<17>::new(
-        uint,
-        ControlBlock::new(
-            d_exp,
-            Sign::Plus,
-            Signals::empty(),
-            Context::default(),
-            ExtraPrecision::new(),
-        ),
-    );
-
-    if b_exp < 0 {
-        let d = D::new(
-            UInt::from_digit(mant),
-            ControlBlock::new(
-                d_exp,
-                sign,
-                Signals::empty(),
-                Context::default(),
-                ExtraPrecision::new(),
-            ),
-        );
-        d.div(psi).transmute()
-    } else {
-        let d = D::new(
-            UInt::from_digit(mant),
-            ControlBlock::new(
-                -d_exp,
-                sign,
-                Signals::empty(),
-                Context::default(),
-                ExtraPrecision::new(),
-            ),
-        );
-        d.mul(psi).transmute()
-    }
-}
 
 macro_rules! from_float_impl {
     ($n: ident, $f: ident) => {
@@ -104,7 +43,7 @@ macro_rules! from_float_impl {
                 // subnormal
 
                 let pow = (MAX_EXP - 2) as i16 + (MANTISSA_DIGITS - 1) as i16;
-                f2dec(frac as u64, -pow, sign)
+                f2dec::f2dec(frac as u64, -pow, sign)
             } else {
                 // normal
 
@@ -133,9 +72,9 @@ macro_rules! from_float_impl {
                     let reduced_frac = frac >> trailing_zeros;
                     let reduced_pow = pow + trailing_zeros as i16;
 
-                    f2dec(reduced_frac as u64, reduced_pow, sign)
+                    f2dec::f2dec(reduced_frac as u64, reduced_pow, sign)
                 } else {
-                    f2dec(frac as u64, pow, sign)
+                    f2dec::f2dec(frac as u64, pow, sign)
                 }
             }
         }