+++ CHANGES.md
@@ -4,6 +4,18 @@ All user-visible changes to this library will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/), as described
 for Rust libraries in [RFC #1105](https://github.com/rust-lang/rfcs/blob/master/text/1105-api-evolution.md)
 
+## [0.2.2] – 2025-02-22
+
+### Changed
+
+- Some float2decimal performance optimizations.
+
+### Fixed
+
+- Minor fixes in float2decimal conversion.
+- Minor fixes with extra precision digits.
+
+
 ## [0.2.1] – 2025-02-18
 
 ### Fixed